+++ documentation/en/SUMMARY.md
@@ -34,9 +34,15 @@
   * [Retrievals](curio-market/retrievals.md)
   * [Migrating From Boost](curio-market/migrating-from-boost.md)
 * [Market 2.0](market-2.0/README.md)
-  * [Products](market-2.0/products.md)
-  * [Market Contracts](market-2.0/contracts.md)
-  * [API](market-2.0/api.md)
+  * [Architecture](market-2.0/architecture.md)
+  * [HTTP API](market-2.0/http-api.md)
+  * [Deal Processing](market-2.0/deal-processing.md)
+  * [Products](market-2.0/products/README.md)
+    * [DDO v1](market-2.0/products/ddo_v1.md)
+    * [Retrieval v1](market-2.0/products/retrieval_v1.md)
+    * [PDP v1](market-2.0/products/pdp_v1.md)
+  * [Contracts](market-2.0/contracts/README.md)
+    * [DDO: CurioDealView v1](market-2.0/contracts/curiodealview.md)
 * [Snap Deals](snap-deals.md)
 * [Batch Sealing](supraseal.md)
 * [Scaling Curio cluster](scaling-curio-cluster.md)
