+++ harmony/harmonydb/downgrade/20260416-mk20-ddo-contracts.sql
@@ -13,6 +13,7 @@ DECLARE
   _headers JSONB;
   _raw_size BIGINT;
   _deal_aggregation INT;
+  _long_term BOOLEAN;
   _piece_id BIGINT;
   _ref_id BIGINT;
 BEGIN
@@ -27,28 +28,32 @@ BEGIN
     END IF;
 
     -- 2. Get deal_aggregation flag
-    SELECT deal_aggregation
-    INTO _deal_aggregation
+    SELECT raw_size, deal_aggregation
+    INTO _raw_size, _deal_aggregation
     FROM market_mk20_pipeline
-    WHERE id = _id AND piece_cid_v2 = _piece_cid_v2 LIMIT 1;
+    WHERE id = _id AND piece_cid_v2 = _piece_cid_v2
+    LIMIT 1;
 
-    -- 3. Look for an existing piece
-    SELECT id
-    INTO _piece_id
-    FROM parked_pieces
-    WHERE piece_cid = _piece_cid AND piece_padded_size = _piece_size;
+    -- 3. Early exit if no deal aggregation match found
+    IF NOT FOUND THEN
+        RETURN FALSE;
+    END IF;
+
+    _long_term := NOT (_deal_aggregation > 0);
 
     -- 4. Insert piece if it is not found
-    IF NOT FOUND THEN
-        INSERT INTO parked_pieces (piece_cid, piece_padded_size, piece_raw_size, long_term)
-        VALUES (_piece_cid, _piece_size, _raw_size, NOT (_deal_aggregation > 0))
-        RETURNING id INTO _piece_id;
-    END IF;
+    INSERT INTO parked_pieces (piece_cid, piece_padded_size, piece_raw_size, long_term)
+    VALUES (_piece_cid, _piece_size, _raw_size, _long_term)
+    ON CONFLICT (piece_cid, piece_padded_size, long_term)
+        WHERE cleanup_task_id IS NULL
+    DO UPDATE
+        SET piece_raw_size = EXCLUDED.piece_raw_size
+    RETURNING id INTO _piece_id;
 
     -- 5. Insert piece ref
     INSERT INTO parked_piece_refs (piece_id, data_url, data_headers, long_term)
-    VALUES (_piece_id, _url, _headers, NOT (_deal_aggregation > 0))
-        RETURNING ref_id INTO _ref_id;
+    VALUES (_piece_id, _url, _headers, _long_term)
+    RETURNING ref_id INTO _ref_id;
 
     -- 6. Insert or update download pipeline with ref_id
     INSERT INTO market_mk20_download_pipeline (id, piece_cid_v2, product, ref_ids)
