+++ market/mk12/mk12.go
@@ -36,10 +36,12 @@ import (
 	"github.com/filecoin-project/curio/harmony/harmonydb"
 	"github.com/filecoin-project/curio/lib/multictladdr"
 	"github.com/filecoin-project/curio/lib/paths"
+	"github.com/filecoin-project/curio/market/backpressure"
 	"github.com/filecoin-project/curio/market/mk12/legacytypes"
 
 	"github.com/filecoin-project/lotus/api"
 	"github.com/filecoin-project/lotus/chain/types"
+	"github.com/filecoin-project/lotus/lib/lazy"
 	"github.com/filecoin-project/lotus/storage/ctladdr"
 )
 
@@ -59,14 +61,15 @@ type MK12API interface {
 }
 
 type MK12 struct {
-	miners     []address.Address
-	db         *harmonydb.DB
-	api        MK12API
-	si         paths.SectorIndex
-	cfg        *config.CurioConfig
-	sm         map[address.Address]abi.SectorSize
-	as         *multictladdr.MultiAddressSelector
-	cidGravity map[address.Address]string
+	miners       []address.Address
+	db           *harmonydb.DB
+	api          MK12API
+	si           paths.SectorIndex
+	cfg          *config.CurioConfig
+	sm           map[address.Address]abi.SectorSize
+	as           *multictladdr.MultiAddressSelector
+	cidGravity   map[address.Address]string
+	backpressure *backpressure.CachedBackPressure
 }
 
 type validationError struct {
@@ -75,7 +78,15 @@ type validationError struct {
 	reason string
 }
 
-func NewMK12Handler(miners []address.Address, db *harmonydb.DB, si paths.SectorIndex, mapi MK12API, cfg *config.CurioConfig, as *multictladdr.MultiAddressSelector) (*MK12, error) {
+func NewMK12Handler(
+	miners []address.Address,
+	db *harmonydb.DB,
+	si paths.SectorIndex,
+	mapi MK12API,
+	cfg *config.CurioConfig,
+	as *multictladdr.MultiAddressSelector,
+	bp *lazy.Lazy[*backpressure.CachedBackPressure],
+) (*MK12, error) {
 	ctx := context.Background()
 
 	sm := make(map[address.Address]abi.SectorSize)
@@ -108,15 +119,21 @@ func NewMK12Handler(miners []address.Address, db *harmonydb.DB, si paths.SectorI
 		}
 	}
 
+	b, err := bp.Val()
+	if err != nil {
+		return nil, xerrors.Errorf("failed to get backpressure: %w", err)
+	}
+
 	return &MK12{
-		miners:     miners,
-		db:         db,
-		api:        mapi,
-		si:         si,
-		sm:         sm,
-		cfg:        cfg,
-		as:         as,
-		cidGravity: cgMap,
+		miners:       miners,
+		db:           db,
+		api:          mapi,
+		si:           si,
+		sm:           sm,
+		cfg:          cfg,
+		as:           as,
+		cidGravity:   cgMap,
+		backpressure: b,
 	}, nil
 }
 
@@ -650,192 +667,16 @@ func (m *MK12) processDeal(ctx context.Context, deal *ProviderDealState) (*Provi
 // Check if WaitDealSectors > MaxQueueDealSector
 // Check for buffered sector at each state of pipeline to their respective Max
 func (m *MK12) maybeApplyBackpressure(ctx context.Context, maddr address.Address) (wait bool, err error) {
-	var totalSize int64
-	err = m.db.QueryRow(ctx, `SELECT COALESCE(SUM(piece_size), 0) AS total_piece_size
-							FROM market_mk12_deal_pipeline
-							WHERE sector IS NULL`).Scan(&totalSize)
+	pressure, err := m.backpressure.MK12Pressure(ctx, &m.cfg.Ingest, m.db)
 	if err != nil {
-		return false, xerrors.Errorf("failed to get cumulative deal size in process from DB: %w", err)
+		return false, xerrors.Errorf("failed to get MK12 pressure: %w", err)
 	}
 
-	maxDsz := m.cfg.Market.StorageMarketConfig.MK12.MaxConcurrentDealSizeGiB >> 30
-	if maxDsz != 0 && totalSize > maxDsz {
-		log.Infow("backpressure", "reason", "too many deals in process", "ConcurrentDealSize", totalSize, "max", m.cfg.Market.StorageMarketConfig.MK12.MaxConcurrentDealSizeGiB)
+	if pressure {
 		return true, nil
 	}
 
-	cfg := m.cfg.Ingest
-
-	// Check market pipeline conditions
-	// We reuse the pipeline stages logic from PipelineStatsMarket to determine
-	// how many pipelines are running and how many are queued at downloading/verify stages.
-	var runningPipelines, downloadingPending, verifyPending int64
-	err = m.db.QueryRow(ctx, `
-WITH pipeline_data AS (
-    SELECT dp.uuid,
-           dp.complete,
-           dp.commp_task_id,
-           dp.psd_task_id,
-           dp.find_deal_task_id,
-           dp.indexing_task_id,
-           dp.sector,
-           dp.after_commp,
-           dp.after_psd,
-           dp.after_find_deal,
-           pp.task_id AS downloading_task_id
-    FROM market_mk12_deal_pipeline dp
-    LEFT JOIN parked_pieces pp ON pp.piece_cid = dp.piece_cid
-    WHERE dp.complete = false
-),
-joined AS (
-    SELECT p.*,
-           dt.owner_id AS downloading_owner,
-           ct.owner_id AS commp_owner,
-           pt.owner_id AS psd_owner,
-           ft.owner_id AS find_deal_owner,
-           it.owner_id AS index_owner
-    FROM pipeline_data p
-    LEFT JOIN harmony_task dt ON dt.id = p.downloading_task_id
-    LEFT JOIN harmony_task ct ON ct.id = p.commp_task_id
-    LEFT JOIN harmony_task pt ON pt.id = p.psd_task_id
-    LEFT JOIN harmony_task ft ON ft.id = p.find_deal_task_id
-    LEFT JOIN harmony_task it ON it.id = p.indexing_task_id
-)
-SELECT
-    COUNT(DISTINCT uuid) FILTER (
-        WHERE (downloading_task_id IS NOT NULL AND downloading_owner IS NOT NULL)
-           OR (commp_task_id IS NOT NULL AND commp_owner IS NOT NULL)
-           OR (psd_task_id IS NOT NULL AND psd_owner IS NOT NULL)
-           OR (find_deal_task_id IS NOT NULL AND find_deal_owner IS NOT NULL)
-           OR (indexing_task_id IS NOT NULL AND index_owner IS NOT NULL)
-    ) AS running_pipelines,
-    COUNT(*) FILTER (WHERE downloading_task_id IS NOT NULL AND downloading_owner IS NULL) AS downloading_pending,
-    COUNT(*) FILTER (WHERE commp_task_id IS NOT NULL AND commp_owner IS NULL) AS verify_pending
-FROM joined
-`).Scan(&runningPipelines, &downloadingPending, &verifyPending)
-	if err != nil {
-		return false, xerrors.Errorf("failed to query market pipeline backpressure stats: %w", err)
-	}
-
-	if cfg.MaxMarketRunningPipelines.Get() != 0 && runningPipelines > int64(cfg.MaxMarketRunningPipelines.Get()) {
-		log.Infow("backpressure", "reason", "too many running market pipelines", "running_pipelines", runningPipelines, "max", cfg.MaxMarketRunningPipelines.Get())
-		return true, nil
-	}
-
-	if cfg.MaxQueueDownload.Get() != 0 && downloadingPending > int64(cfg.MaxQueueDownload.Get()) {
-		log.Infow("backpressure", "reason", "too many pending downloads", "pending_downloads", downloadingPending, "max", cfg.MaxQueueDownload.Get())
-		return true, nil
-	}
-
-	if cfg.MaxQueueCommP.Get() != 0 && verifyPending > int64(cfg.MaxQueueCommP.Get()) {
-		log.Infow("backpressure", "reason", "too many pending CommP tasks", "pending_commp", verifyPending, "max", cfg.MaxQueueCommP.Get())
-		return true, nil
-	}
-
-	// Existing logic for snap vs porep pipelines
-	if cfg.DoSnap {
-		var bufferedEncode, bufferedProve, waitDealSectors int
-		err = m.db.QueryRow(ctx, `
-		WITH BufferedEncode AS (
-			SELECT COUNT(p.task_id_encode) - COUNT(t.owner_id) AS buffered_encode
-			FROM sectors_snap_pipeline p
-			LEFT JOIN harmony_task t ON p.task_id_encode = t.id
-			WHERE p.after_encode = false
-		),
-		 BufferedProve AS (
-			 SELECT COUNT(p.task_id_prove) - COUNT(t.owner_id) AS buffered_prove
-			 FROM sectors_snap_pipeline p
-			 LEFT JOIN harmony_task t ON p.task_id_prove = t.id
-			 WHERE p.after_prove = true AND p.after_move_storage = false
-		 ),
-		 WaitDealSectors AS (
-			SELECT COUNT(DISTINCT osp.sector_number) AS wait_deal_sectors_count
-			FROM open_sector_pieces osp
-			LEFT JOIN sectors_snap_initial_pieces sip 
-				 ON osp.sector_number = sip.sector_number
-			WHERE sip.sector_number IS NULL
-		 )
-		SELECT
-			(SELECT buffered_encode FROM BufferedEncode) AS total_encode,
-			(SELECT buffered_prove FROM BufferedProve) AS buffered_prove,
-			(SELECT wait_deal_sectors_count FROM WaitDealSectors) AS wait_deal_sectors_count
-		`).Scan(&bufferedEncode, &bufferedProve, &waitDealSectors)
-		if err != nil {
-			return false, xerrors.Errorf("counting buffered sectors: %w", err)
-		}
-
-		if cfg.MaxQueueDealSector.Get() != 0 && waitDealSectors > cfg.MaxQueueDealSector.Get() {
-			log.Infow("backpressure", "reason", "too many wait deal sectors", "wait_deal_sectors", waitDealSectors, "max", cfg.MaxQueueDealSector.Get())
-			return true, nil
-		}
-
-		if cfg.MaxQueueSnapEncode.Get() != 0 && bufferedEncode > cfg.MaxQueueSnapEncode.Get() {
-			log.Infow("backpressure", "reason", "too many encode tasks", "buffered", bufferedEncode, "max", cfg.MaxQueueSnapEncode.Get())
-			return true, nil
-		}
-
-		if cfg.MaxQueueSnapProve.Get() != 0 && bufferedProve > cfg.MaxQueueSnapProve.Get() {
-			log.Infow("backpressure", "reason", "too many prove tasks", "buffered", bufferedProve, "max", cfg.MaxQueueSnapProve.Get())
-			return true, nil
-		}
-	} else {
-		var bufferedSDR, bufferedTrees, bufferedPoRep, waitDealSectors int
-		err = m.db.QueryRow(ctx, `
-		WITH BufferedSDR AS (
-			SELECT COUNT(p.task_id_sdr) - COUNT(t.owner_id) AS buffered_sdr_count
-			FROM sectors_sdr_pipeline p
-			LEFT JOIN harmony_task t ON p.task_id_sdr = t.id
-			WHERE p.after_sdr = false
-		),
-		BufferedTrees AS (
-			SELECT COUNT(p.task_id_tree_r) - COUNT(t.owner_id) AS buffered_trees_count
-			FROM sectors_sdr_pipeline p
-			LEFT JOIN harmony_task t ON p.task_id_tree_r = t.id
-			WHERE p.after_sdr = true AND p.after_tree_r = false
-		),
-		BufferedPoRep AS (
-			SELECT COUNT(p.task_id_porep) - COUNT(t.owner_id) AS buffered_porep_count
-			FROM sectors_sdr_pipeline p
-			LEFT JOIN harmony_task t ON p.task_id_porep = t.id
-			WHERE p.after_tree_r = true AND p.after_porep = false
-		),
-		WaitDealSectors AS (
-			SELECT COUNT(DISTINCT osp.sector_number) AS wait_deal_sectors_count
-			FROM open_sector_pieces osp
-			LEFT JOIN sectors_sdr_initial_pieces sip 
-				 ON osp.sector_number = sip.sector_number
-			WHERE sip.sector_number IS NULL
-		)
-		SELECT
-			(SELECT buffered_sdr_count FROM BufferedSDR) AS total_buffered_sdr,
-			(SELECT buffered_trees_count FROM BufferedTrees) AS buffered_trees_count,
-			(SELECT buffered_porep_count FROM BufferedPoRep) AS buffered_porep_count,
-			(SELECT wait_deal_sectors_count FROM WaitDealSectors) AS wait_deal_sectors_count
-		`).Scan(&bufferedSDR, &bufferedTrees, &bufferedPoRep, &waitDealSectors)
-		if err != nil {
-			return false, xerrors.Errorf("counting buffered sectors: %w", err)
-		}
-
-		if cfg.MaxQueueDealSector.Get() != 0 && waitDealSectors > cfg.MaxQueueDealSector.Get() {
-			log.Infow("backpressure", "reason", "too many wait deal sectors", "wait_deal_sectors", waitDealSectors, "max", cfg.MaxQueueDealSector.Get())
-			return true, nil
-		}
-
-		if bufferedSDR > cfg.MaxQueueSDR.Get() {
-			log.Infow("backpressure", "reason", "too many SDR tasks", "buffered", bufferedSDR, "max", cfg.MaxQueueSDR.Get())
-			return true, nil
-		}
-		if cfg.MaxQueueTrees.Get() != 0 && bufferedTrees > cfg.MaxQueueTrees.Get() {
-			log.Infow("backpressure", "reason", "too many tree tasks", "buffered", bufferedTrees, "max", cfg.MaxQueueTrees.Get())
-			return true, nil
-		}
-		if cfg.MaxQueuePoRep.Get() != 0 && bufferedPoRep > cfg.MaxQueuePoRep.Get() {
-			log.Infow("backpressure", "reason", "too many PoRep tasks", "buffered", bufferedPoRep, "max", cfg.MaxQueuePoRep.Get())
-			return true, nil
-		}
-	}
-
-	return false, nil
+	return m.backpressure.SectorPressure(ctx, &m.cfg.Ingest, m.db)
 }
 
 // applyFilters is used to validate deal proposals against client and pricing filters
