+++ market/mk20/contract/CurioDealViewV1.abi
@@ -1,87 +1,111 @@
 [
-  {
-    "type": "function",
-    "name": "getDeal",
-    "inputs": [
-      {
-        "name": "dealId",
-        "type": "uint256",
-        "internalType": "uint256"
-      }
-    ],
-    "outputs": [
-      {
-        "name": "",
-        "type": "tuple",
-        "internalType": "struct ICurioDealViewV1.DealView",
-        "components": [
-          {
-            "name": "state",
-            "type": "uint8",
-            "internalType": "enum ICurioDealViewV1.DealState"
-          },
-          {
-            "name": "providerActorId",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "clientId",
-            "type": "bytes",
-            "internalType": "bytes"
-          },
-          {
-            "name": "pieceCidV2",
-            "type": "bytes",
-            "internalType": "bytes"
-          },
-          {
-            "name": "startEpoch",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "duration",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "allocationId",
-            "type": "uint256",
-            "internalType": "uint256"
-          },
-          {
-            "name": "finalizedEpoch",
-            "type": "uint256",
-            "internalType": "uint256"
-          }
-        ]
-      }
-    ],
-    "stateMutability": "view"
-  },
-  {
-    "type": "function",
-    "name": "version",
-    "inputs": [],
-    "outputs": [
-      {
-        "name": "",
-        "type": "uint256",
-        "internalType": "uint256"
-      }
-    ],
-    "stateMutability": "pure"
-  },
-  {
-    "type": "error",
-    "name": "DealNotFound",
-    "inputs": [
-      {
-        "name": "dealId",
-        "type": "uint256",
-        "internalType": "uint256"
-      }
-    ]
-  }
+    {
+        "inputs": [
+            {
+                "internalType": "uint256",
+                "name": "dealId",
+                "type": "uint256"
+            }
+        ],
+        "name": "DealNotFound",
+        "type": "error"
+    },
+    {
+        "inputs": [
+            {
+                "components": [
+                    {
+                        "internalType": "uint256",
+                        "name": "dealId",
+                        "type": "uint256"
+                    },
+                    {
+                        "internalType": "enum ICurioDealViewV1.DealState",
+                        "name": "state",
+                        "type": "uint8"
+                    },
+                    {
+                        "internalType": "uint256",
+                        "name": "providerActorId",
+                        "type": "uint256"
+                    },
+                    {
+                        "internalType": "bytes",
+                        "name": "clientId",
+                        "type": "bytes"
+                    },
+                    {
+                        "internalType": "bytes",
+                        "name": "pieceCidV2",
+                        "type": "bytes"
+                    },
+                    {
+                        "internalType": "uint256",
+                        "name": "startEpoch",
+                        "type": "uint256"
+                    },
+                    {
+                        "internalType": "uint256",
+                        "name": "duration",
+                        "type": "uint256"
+                    },
+                    {
+                        "internalType": "uint256",
+                        "name": "allocationId",
+                        "type": "uint256"
+                    },
+                    {
+                        "internalType": "uint256",
+                        "name": "finalizedEpoch",
+                        "type": "uint256"
+                    }
+                ],
+                "internalType": "struct ICurioDealViewV1.CurioDealView",
+                "name": "deal",
+                "type": "tuple"
+            }
+        ],
+        "name": "verifyDeal",
+        "outputs": [
+            {
+                "internalType": "bool",
+                "name": "",
+                "type": "bool"
+            }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+    },
+    {
+        "inputs": [],
+        "name": "version",
+        "outputs": [
+            {
+                "internalType": "uint256",
+                "name": "",
+                "type": "uint256"
+            }
+        ],
+        "stateMutability": "pure",
+        "type": "function"
+    },
+    {
+        "inputs": [
+            {
+                "internalType": "uint256",
+                "name": "dealId",
+                "type": "uint256"
+            }
+        ],
+        "name": "getDealState",
+        "outputs": [
+            {
+                "internalType": "enum ICurioDealViewV1.DealState",
+                "name": "",
+                "type": "uint8"
+            }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+    }
 ]
