+++ market/mk20/contract/utils.go
@@ -1,0 +1,7 @@
+package contract
+
+const (
+	DealStatusOpen uint8 = iota
+	DealStatusActive
+	DealStatusFinalized
+)
