+++ market/mk20/http/swagger.json
@@ -916,6 +916,10 @@
                             "$ref": "#/definitions/address.Address"
                         }
                     ]
+                },
+                "start_epoch": {
+                    "description": "StartEpoch optionally specifies the epoch by which a deal should be active on the chain",
+                    "type": "integer"
                 }
             }
         },
